ENH: Add context menu when right-clicking in slice or 3D view
Currently, the only actions in the view context menu (when clicking on empty area) to change interaction mode and copy view content to clipboard.

Also added menu sections and ordering. "section" property in the menu action can be set using qSlicerSubjectHierarchyAbstractPlugin::setActionPosition helper function.
All actions that are in the same section value after rounding to integer are in the same section. The fractional part of the value can be used to order the actions (items with lower values appear higher in the menu).