ENH: Add interface to change default sequence storage nodes
Currently, it is not possible in a module to change the default SequenceStorageNode for a data node class in the Slicer core.
This commit adds an interface so that Extensions can now add default sequence storage node classes for specific data types by registering the data node/storage node in the default data node.

Example of use:
````
  streamingVolumeNode = slicer.mrmlScene.CreateNodeByClass("vtkMRMLStreamingVolumeNode")
  streamingVolumeNode.SetDefaultSequenceStorageNodeClassName("vtkMRMLStreamingVolumeSequenceStorageNode")
  slicer.mrmlScene.AddDefaultNode(streamingVolumeNode)
  streamingVolumeNode.UnRegister(None)
````