BUG: Fix coloring of model exported from segmentation
A user reported that when a model exported from segmentation is loaded into Slicer then it appears in purple color and its color cannot be changed.
This was due to ImageScalars point array being present in the exported model that is displayed by default when a model is loaded into the scene.

Fixed by not removing ImageScalars array from the exported model.

Also implemented automatic disabling of scalars display when a solid color is set in the subject hierarchy tree or models module to make sure that when the user changes the color then it actually takes effect.