ENH: Only show Edit properties in context menu when editing is available
It was confusing for users that "Edit properties..." menu was shown when actually there was no specific module associated with the node.

It also caused a crash in Segment Editor's Surface cut module when the user clicked on "Edit properties..." on a node that was not intended to be edited by users (it was set to HiddenFromEditors).
The problem was that editing a node triggered a module switch, and the Surface cut effect deleted the node in response to module change (as the node was just meant to be used temporarily).