ENH: "Apply to all segments" CB added
Added a new checkbox and logic that enables serial smoothing of all visible segments currently open in the Segment Editor. If no segments are visible there is a logging message. The last selected segment gets restored.

The progress is shown as a status message.

Some more Minor fixes.
MMLabel changed.