ENH: Update scalar volume display to use full color node scalar range
Within vtkMRMLScalarVolumeDisplayNode, scalar values are mapped using window/level to a unsigned char scalar range of 0 to 255.
While most color nodes in the scene use a range of 0 to 255, some nodes use different ranges (ex. RedGreenBlue uses -6.0 to 6.0).

In order to make full use of all colors in the table, we now copy the color node lookup table, and map it to the expected 0 to 255 scalar range from vtkImageMapToWindowLevelColors.

This commit also adds support for using the scalar range flag of UseDirectMapping to map the values of the image directly onto the scalar range of the table, bypassing the window/level.