BUG: Fix hang when closing scene (#5745)
Problem is that on scene close in certain cases, when the new SH node is created, the Markups displayable manager requests render on the node added event of the SH node, then there is an endless loop in the Folder plugin's GetHierarchyVisibility because SH item parent is 0 (i.e. invalid ID) and never reaches the scene.