ENH: Report more scene loading errors
If a node fails to be instantiated then a message is displayed to the user.

List of installed extensions is saved in the scene and in case a scene loading error occurs then the list of missing extensions (that were installed at the time of saving the scene but not now) are displayed to the user. Since most extensions do not affect scene loading, if some extensions are not installed but scene loading is successful then no warning is displayed.