BUG: Fix Drawing effect completion by right-click
Right-click to finish Draw effect in Segment Editor made the view context menu displayed.
Fixed this by making the effect consume the right-click event.

Also added the option to finish the contour with double-click. Double-click was not used before because this even has just recently been added to VTK.

re #5734