BUG: Save segmentation imported from plain image at original geometry
When loading a segmentation from a plain image (rather than seg.nrrd), the reference image geometry of the original image was not saved.
Due to the missing geometry conversion parameter, exporting/saving the segmentation would cause the segmentation to be exported/saved at minimum extent. Fixed by setting the reference image geometry conversion parameter to the size of the original image.

Re #5323