BUG: Force update of grow from seeds 3D closed surface preview
Due to a previous optimization that set the labelmap representation of the preview segments directly, the 3D preview closed surface was no longer automatically updating. To fix this, we remove the existing closed surface so that it will be reconverted after the segmentation is updated.

When the functions for setting multiple segments are added, we will no longer need this call as the closed surface update will be handled by the segmentation logic.

Re #5757