BUG: Fix invalid overlap when splitting segments
When modifying a segment the list of segments that need to be overwritten does not normally include segments that are on the same layer (since they will be overwritten by default). This list was populated before the selected segment is modified. This causes segments on the same layer to not be overwritten if the selected segment is separated to a new layer.

Fixed by populating the list of segments to overwrite after the selected segment has been modified (and potentially moved to a new layer).

Re #5338