BUG: Prevent resetting of plot series node names
Checked state of plot series node selector widget was always temporarily reset when a plot series node was changed.
This change triggered MRML update from the GUI when a plot series name was changed, thereby immediately reverting the node change.

Fixed the issue by only updating the checked state in the widget if it is actually changed.

An even more robust solution would be to only update the MRML node name from the scene model item if the name is changed in the GUI.
However, this is hard because in the checkable combobox the checkbox is not a separate item but it belongs to the name item.