ENH: Make DICOM retrieve more efficient by optimizing indexing
Indexing is now performed in the background and database is updated (waitForImportFinished() is called and already indexed files are removed) periodically, after many files are imported, not after each file.
It is important _not_ to call waitForImportFinished() after each file because the method has an overhead that increases as the database size grows (it has to update displayed fields).

Database update is forced when no new file is received for a while (detected using a single-shot timer) or number of unprocessed incoming files reaches a certain limit.

fixes #5773