ENH: Incorporate ExportAs module into Slicer core
See #4989

ExportAs is a Slicer module that was created for the extension SlicerMorph.
This commit incorporates ExportAs into Slicer core, renaming the module from
`ExportAs` to `DataSubjectHierarchyPlugin`.

This subject hierarchy plugin is associated with the Data module as a whole,
rather than with any specific module. This is why I am organizing it as
a `SubjectHierarchyPlugin` under `Modules/Loadable/Data`.

I followed `Modules/Loadable/Annotations/SubjectHierarchyPlugins/`
as a model for how to set this up, since it is an example of a subject hierarchy
plugin that is written in python and added under a loadable module.