ENH: Improve Segment Editor widget layout
Effect list is displayed on the left in a vertical grid, icon-only. This takes less space, especially less vertical space, therefore less scrolling is needed.

The motivation for this change is to better use space in the Segment Editor widget to improve usability, as discussed at https://discourse.slicer.org/t/segment-editor-effect-button-size-consistency/19214.

Co-authored-by: Andras Lasso <lasso@queensu.ca>