ENH: Allow temporarily maximize a view in the layout
Clicking "maximize" button in the header of the view controller makes the corresponding view maximized in the view layout. Clicking the button again restores the original view layout.

Shortcuts added to access this feature without leaving the view:
- double-click in an empty area in the view maximizes/restores a view
- "Maximize" / "Restore view layout" actions view context menu (right-clicking in an empty area)

Implementation: A new MaximizedViewNode node reference is added to vtkMRMLLayoutNode. If MaximizedViewNode is set to a valid view node then that view is displayed instead of the views defined by vtkMRMLLayoutNode::GetCurrentLayoutDescription(). The current layout ID or description in the layout node is not changed, the override happens at lower level, in qMRMLLayoutManager. The commit moves storage and observation of view nodes from various view controller widgets to the nodes into the parent qMRMLViewControllerBar to reduce code duplication.

fixes #1409

Co-authored-by: jahnavi0801 <jahnavi0801@gmail.com>