BUG: Remove unnecessary undo state creation in Segment Editor Paint effect in 3D
The Paint effect in Segment Editor does not start painting (just lets other widgets process events) when the use clicks-and-drags over an empty area.
In this cases just the view rotates and no painting occurs, yet an undo state was saved. This commit fixes this: no undo state is created anymore if the user just rotates the view.

fixes #5808