BUG: Fix crash when erasing regions outside segmentation
If the modifier labelmap and resampled segmentation labelmaps do not overlap, then GetExtentIntersection will generate an invalid extent. Passing this extent to vtkImageThreshold::UpdateExtent will cause a crash.

Fixed by not applying the segment mask to the modifier if the extent is invalid. This should not cause any issues since if there is no overlap between the modifier/segment labelmaps, nothing should need to be erased.

Fix: #5812