ENH: Remove DataProbe show zoomed slice
From @lassoan
"Tthe "Show zoomed slice" feature can be removed. It is useless, as it does not show any of the overlays (markups, segmentations, model intersections, etc.), it does not slow down the mouse cursor (so it is very hard to position the mouse accurately), and when you drag-and-drop markup points then the crosshair position is not the markup control position. Simply zooming in the image, or using the Magnifier feature built-into the operating system both work much better."