ENH: Always restore maximized view when selecting a layout on the toolbar
Addresses this request from @jamesobutler: "if I maximize the red slice of the Four-Up view, then accidentally forget my original layout is Four-Up and then choose Four-Up in the layout toolbar selector (yes, the toolbar icon will reflect Four-up, but might not notice this), should that restore it as well? Currently it does nothing I'm assuming because it detects that there is no layout change to process."

Now if a view layout is selected it always sets the regular (non-maximized) layout, even if the current layout is chosen again.

Also improved resetting of the layout. When scene is closed then now the view layout is restored from maximized state.