BUG: Fix disappearing icons in color node selector
When the scene model was rebuilt from scratch (for example, due to batch update or closing the scene), the color gradient icons disappeared from the color node selector qMRMLColorTableComboBox.

The problem was that the icons were lost because the model items were recreated, but the nodes were still the same (singleton nodes are not deleted when the scene is closed), so the icons were not set again.

fixes #5834
fixes #4721