ENH: Left/right dock widgets given ownership to bottom corner areas
This is primarily motivated to give the module panel more dedicated vertical space by default.

Since the module panel is typically not a majority of the width dimension, this means the python interactor in the bottom widget area still has a wide aspect ratio. The Python interactor can be still docked to the top for full-width display.

Since QDockWidget positioning is saved as part of the window state, this change won't populate to users if they already have a Slicer.ini. They would have to restore defaults in settings or manually do the following code to trigger the newly defined state.