ENH: Improve logic of subject hierarchy volume rendering display
When in the subject hierarchy tree's view menu "Show in 3D views as volume rendering" was enabled once, then volume rendering kept reappearing when the eye icon was clicked (because if volume rendering display node was found then the volume got always displayed in the 3D views).

Now when the user disables "Show in 3D views as volume rendering" in the view menu then volume rendering does not get re-enabled again by clicking the eye icon (only when user explicitly enables "Show in 3D views as volume rendering" in the view menu).