BUG: Fix crash by disabling drag&drop in SH combobox
In Subject Hierarchy combobox when the user started a drag, a crash occurred. Since comboboxes are intended only for selection of items and not reorganizing them, and even if we want to drag&drop it is quite inconvenient and hard to control, it makes sense to simply disable the drag&drop feature.