BUG: Ensure that masking settings are fully observed for most effects
The current region previously always added the current segment to the editable area if we were using one of the "EditAllowedInside*Segment*" masking modes.

This commit changes the behavior to only remove the current segment from the mask when:
  1) The masking setting is set to inside a single segment
  2) Modification mode is ModificationModeRemove

In this circumstance, we add the "Mask" segment to the current segment.

Re #5831