ENH: Build SlicerApp-real.exe as a console application to enable console output
If SlicerApp-real.exe is built as a console application then it is possible to capture and redirect its output (to files, display, any other software), and test outputs also become visible on dashboard builds.

If the launcher is still built as a Windows GUI application then no command console is displayed when Slicer is started.

Added instructions (in Docs\developer_guide\debugging\overview.md) for accessing the console output when launcher is a Windows GUI application.

re #2934