ENH: Build SlicerApp-real.exe as a console application to enable console output (corrected)
If SlicerApp-real.exe is built as a console application then it is possible to capture and redirect its output (to files, display, any other software), and test outputs also become visible on dashboard builds.

If the launcher is still built as a Windows GUI application then no command console is displayed when Slicer is started.

Added instructions (in Docs\developer_guide\debugging\overview.md) for accessing the console output when launcher is a Windows GUI application.

This is a corrected version of a previous commit that was reverted because it contained an error (in SlicerMacroBuildApplication.cmake, the condition to set SLICERAPP_EXE_OPTIONS to WIN32 was based on Slicer_HAS_CONSOLE_LAUNCHER instead of Slicer_HAS_CONSOLE_IO_SUPPORT).

re #2934