ENH: Add method to specify allowed view context menu actions for specific nodes
Until now, view context menu actions (that are displayed when right-clicking in views) were only configurable globally, for all nodes.
This commits add further filtering for specific subject hierarchy items.

This is useful for example when a module creates a temporary markup node to get input points from a user and the user should not be allowed to delete that node.