BUG: Fix Segment Editor undo history corruption when adding a segment without status tag
When a segment was added to an edited segmentation that did not contain a status tag then that segment was modified and then restored with the undo history, the undo history got cleared.
It was due to that the originally empty status tag got replaced by Segmentation.Status=notstarted, which triggered a modified event, which cleared all redo states.

fixes #5566