COMP: Simplify extension package upload logic removing midas support
This commit is similar to 64741672b (COMP: Simplify application package
upload logic removing midas support) expect it applies to extension
packages.

It removes support to upload to the deprecated midas based
server and makes the upload to the Girder based server required ensuring
it reports an error in case of upload failure.