BUG: Corrects fixed point number property on loading and minor issues
This commit fixes a bug causing the fixed point number property to not be read while loading a markup node from file. This bug is documented in issue #5889

Other minor issues with the Markups toolbar addressed in this commit are: changing text on "Display markups toolbar" button to "Toggle markups toolbar", correcting size of node selector and color picking button, and adding a space between the create node buttons and the node selector.