BUG: Fix markups place widget not able to switch active node
When clicking place button in a markups place widget, the active markups node was not changed.
It was due to modifying the selection node in qSlicerMarkupsPlaceWidget::updateWidget().

Fixed by moving SelectionNode->SetActivePlaceNodePlacementValid() calls to the markups logic.

Also cleaned up and simplified the code and fixed indentation issues at a couple of places.