ENH: Remove vtkMRMLMarkupsROINode::GetBoundsROI() and add documentation
The GetBoundsROI function was confusingly named and did not do what users expected (returning the bounds of the ROI in node coordinates).
This commit removes the GetBoundsROI function, and adds additional documentation to explain which functions can be used to determine the ROI axis-aligned bounding box and oriented bounding box.