BUG: Fixed placed markup points deletion on view layout change
In some cases, a view layout change deleted a previously placed control point.

It was caused by an incorrect check that decided that a control point should be removed when the mouse leaves a view.
The problem was fixed by making sure that the control point was only removed if it was in preview state.