ENH: Improve control point position status options for clarity
Position status actions have been reordered and renamed in the right-click context menu of the control point table. Position status icons were also added to the actions.

"Clear" is chosen over "Unset" since it is less technical. Action order is based on the cycle order when click in the position status column of the markups control point table.