ENH: Make segment effect cursor hotspot position clearer
The "Draw" effect cursor was often mistaken for having the hotspot positioned at the end of the Pencil rather than the end of the blue arrow in the cursor.  The blue arrow has been replaced with the standard white arrow cursor arrow and a white border has been added around the effect image to make it clearer that the hotspot position is not in the effect image.