ENH: Improved DICOM cine-MRI loading logic
Cine-MRI frames can be loaded as a sequence of frames (by DICOMImageSequencePlugin) or a sequence of volumes (by MultiVolumeImporterPlugin).
If the series contain frames with different orientations then a sequence of frames should be loaded (and later reconstructed into volume sequence for example using SlicerHeart extension's Reconstruct 4D MRI module).
If the series contain frame in the same orientation then it is better to give multivolume importer a chance to load it as a volume sequence.