ENH: Add option to suppress confirm popups during extension installation
When an extension is installed that depends on other extensions then a popup is displayed for the user to confirm installation of additional extensions.
When installing extensions from a script or Jupyter notebooks, these popups may block the application.

Added an "interactive" property (enabled by default) to qSlicerExtensionsManagerModel that allows disabling these popups.