BUG: Fix markups curve point projection algorithm
When the code was refactored, a bug was introduced in vtkProjectMarkupsCurvePointsFilter::PointProjectionHelper::GetPointNormals.

The `double* GetPoint(n)` method uses a single temporary double[3] array to return point coordinates herefore if coordinates of multiple
points need to be retrieved then the `void GetPoint(vtkIdType id, double x[3])` must be used instead.