ENH: Improve step size in segmentation display widget
Reduced step sizes for all percentage sliders in the segmentation display widget to 0.1 page step and 0.05 single step, as per
https://discourse.slicer.org/t/segmentations-module-clicking-on-slider-track-should-move-slider-in-steps-spinner-should-increment-last-digit/20376/2

"Page size is typically 30% in most of the percentage sliders, but I agree that 10% page step would be a bit more useful.
Step size in PowerPoint is set so that you can get from 0% to 100% using the up/down spin sliders in about 6 seconds. This would mean about 5% single step size in Slicer."