BUG: Fix camera translation in 3D views when camera is scaled
Before this change, user interaction (especially dragging) was not scaled correctly when camera was scaled (aspect ratio is changed).
For example, if the camera was stretched out along X direction then translation along X was slower.

Scaling of camera is important in some areas (like seismic exploration) when different axes represent different units (distance and time for example).

This commit fixes this and now user feel comfortable when the camera is scaled. A C++ test was added to ensure that this behavior is preserved even when code will change in the future.