BUG: Fix error in DICOM module when it is used as startup module
When DICOM module was used as the startup (home) module, which Slicer activated on startup then initialization was failed.
The problem was that some initialization steps were performed when `startupCompleted` signal was emitted, which happens after the startup module's widget is set up.

Solved the issue by making sure that all initialization steps are performed in time by calling it from the widget setup method as well.