ENH: Add API to open an URL while the application is running
All unrecognized command-line arguments have been attempted to be interpreted as URLs at application startup (by the application object emitting a urlReceived signal).
But there has been no way to trigger this feature after the application started.

This commit adds an `openUrl` method that allows opening URLs while Slicer is already running.

This is useful for example when Slicer runs as a service because it is much faster to close the scene and load data from a URL rather than restart the application with that new URL.