COMP: Fix VTK_USE_FILE warning when building extensions
When an extension was configured, VTK printed the warning message: "The `VTK_USE_FILE` is no longer used starting with 8.90."
This is due to that in VTK9 defines VTK_USE_FILE variable (so Slicer attempts to use it), but the variable just points to a file (vtk-use-file-deprecated.cmake) that prints the message that the file is deprecated.

Fixed by explicitly disabling usage of VTK_USE_FILE for VTK9.