BUG: Fix place control point button menu button always visible
The Place control point QToolButton was changed in https://github.com/Slicer/Slicer/commit/09d12ae6f7598fe8490b691d17a35dcca5bd344a to use the QToolButton::MenuButtonPopup type for ToolButtonPopupMode. Previously it was the default DelayedPopup. DelayedPopup would hide the menu arrow when there was no menu, however the MenuButtonPopup always shows the button even when there is no menu.

To avoid having a menu button with no menu, we can switch to the DelayedPopup mode when we do not have a menu for the toolbutton.