BUG: Fix erase when using the selected segment as mask
Erasing a single segment would "add-back" the mask segment. When the selected segment and mask segment were the same, this would result in nothing changing. Fixed by only adding the mask segment back if it is not the selected segment.

See discussion: https://discourse.slicer.org/t/erase-using-a-segment-as-only-editable-area/20519

Also adds tests to SegmentationsModuleTest2.py to detect future regressions in masking behavior.