BUG: Fix Segment table update for Segmentation replay in Sequences
When vtkMRMLSegmentationNode were copied from recorded sequences, no events were emitted that would cause the table to update. This commit adds SegmentationChangedEvent to vtkMRMLSegmentationNode, which is triggered whenever the vtkSegmentation object is changed. If the event is observed by qMRMLSegmentsModel, it will trigger a reset of the model.

This commit also makes some changes to qMRMLSegmentsTableView and qMRMLSegmentEditorWidget to ensure that the widget states are consistent during Segmentation node Sequence replay.

Fixes #5953