ENH: Remove Slicer3 Fiducial List classes
vtkMRMLFiducialListNode was originally marked as deprecated in June 2012 (https://github.com/Slicer/Slicer/commit/14d653ae3bf0b861a1ac92941c0a874b5380cace) which is over 9 years ago. It was stated that it was for backwards compatibility for Slicer 3 fiducial lists and that the Annotation module was to be used instead.