ENH: Clean up markups node API
Deprecate all method names for that either of these applies:
- use old terms: fiducial or markup (instead of control point)
- include "array", "vector", "pointer" in the name: this is redundant, as input arguments can be deduced from the type, and return type for vectors is almost always vtkVectorNd (as it is safe, and simple to use)

For methods that return a 3D vector, add a Python friendly variant that returns a vtkVector3d. This allows the user to get a vector in one line instead of two lines (first line creates a vector, second line sets its value).