BUG: Fix intensity masking for Islands effect
Segment editor effects that use "ModificationModeSet" restore the existing segment outside of the masking region. The mask was not being applied, and so the entire segment was restored. Fixed by only restoring the segment outside the mask area.

This commit also adds tests for intensity masking with island effects, and fixes issues that caused the test to fail when run with a main window. (Uninitialized labelmaps + unhandled VTK_SIGNED_CHAR data type in vtkImageLabelOutline).

Fixes #6043