ENH: Color legend rework
Color legend (formerly "color bar" or "scalar bar"), which can display scalar (or label) mapping to colors was completely reworked to conform with Slicer architecture. The new implementation is simpler (does not require manual adjustment of number of colors and scalar range, accessible in subject hierarchy and Colors, Volumes, and Markups modules) and safer (less chance for displayed and actual scalar range getting out of sync), and more capable (multiple color legends can be displayed).

A new display node, vtkMRMLColorLegendDisplayNode was added, which is displayed by vtkMRMLColorLegendDisplayableManager. The old vtkMRMLScalarBarDisplayableManager remained unchanged (including the name, which is now not very accurate) for now, for backward compatibility.

A subject hierarchy plugin is added, which adds the "Show color legend" checkbox to the visibility menu (right-click on eye icon in the data tree) to show-hide color legend with default options. Color legend properties can be edited for all nodes in Color module. A Color legend section was added to Volumes, Models, and Markups modules to allow editing of color legend properties without leaving those modules. qMRMLColorLegendDisplayNodeWidget was added to allow editing of color legend properties in custom GUIs.

Additional improvements:
- Unified design of handling of "additional" display nodes (such as volume rendering and color legend display nodes) that should not be automatically displayed when user request show of the displayable node (e.g., clicking on the eye icon). If the new ShowMode property of the display node is be set to ShowIgnore then show/hide requests of the displayable node will not affect visibility of the display node. This is useful because often color legend or volume rendering is disabled explicitly (using the subject hierarchy visibility menu) to prevent clutter in viewers and in these cases the user would not want to show these again when overall visibility is toggled.
- Added Qt designer plugins for Volumes widgets.
- Improved Color module widget layout: If Discrete and Continuous color lookup table are showed in the same collapsible section. This way if the section is collapsed then it remains collapsed regardless of what color node is selected (so the Color legend section position on screen remains the same).
- Improved Volumes module widget layout: Display section used a QStackedWidget-based widget, which always filled the available space, pushing down the Color legend section. Switched to using a simple QWidget and switch between qSlicer*DisplayWidget widgets manually.

Co-authored-by: Andras Lasso <lasso@queensu.ca>