BUG: Prevent data probe from expanding horizontally
By setting the the horizontal size policy of the data probe widget to "Ignored", the widget will use as much space as is provided horizontally, but will never cause the module panel to change size.
The widget can still expand vertically, allowing strings that are too long to be displayed in a single line to be wrapped and not truncated.