ENH: Always show coordinates in the control points table and clean up updates
The control point list behavior is changed so that point coordinates are displayed even when point is undefined (just grayed out).

The commit also cleans up table updates: new table items are created only if needed, making the code cleaner and updates faster (when there are tens of thousands of points then it can make a big difference if items are recreated or updated).