BUG: Fix "child widget is not accessible" warnings in Markups module
Due to a Qt bug (https://bugreports.qt.io/browse/QTBUG-83327) QTableWidget that is created in a disabled group logs many "child widget is not accessible" warnings in debug mode.
This caused warnings appear in the control point list table in Markups module.

As a workaround (until the Qt bug is fixed), controlPointsCollapsibleButton is disabled in the widget constructor instead in the .ui file.

fixes https://github.com/Slicer/Slicer/issues/5913