ENH: Use markup type internal name when display name not specified
A custom markup "GridSurface" did not specify a GetMarkupTypeDisplayName, so in Slicer it was shown as "Markup" in the Markups module. This isn't helpful so it should fallback to use the internal markup type name of "GridSurface" even though it should have a GetMarkupTypeDisplayName method that defines it as "Grid Surface".