ENH: Add vtkMRMLMarkupsDisplayNode interaction handle scale property
Previously the size of the interaction handles was controlled using the vtkMRMLMarkupsDisplayNode::GlyphScale/GlyphSize properties.

This commit adds the vtkMRMLMarkupsDisplayNode::InteractionHandleScale property, which behaves in the same way as GlyphScale except only affects the interaction handle. The interaction handles are no longer affected by Glyph properties.

Also cleans up the constants used to define the size of the interaction widget.

see #5061