ENH: Add InsideOut flag to markups ROI
The inside out function can be used to invert the ROI so that points inside the ROI are considered outside, and vice versa. This state is reflected in GetPlanes()/GetPlanesWorld() and IsPointInROI()/IsPointInROIWorld().

This commit also adds implicit functions to the ROI node that can be used by other nodes, and that will invert their sign based on the InsideOut state.

Modules can choose to use this flag any way it makes sense in that context. In the future, the flag can be used for changing the appearance of the ROI (for example, the ROI can be filled inside or outside).

see #5061