BUG: Fix selection restoration issue in SH tree view
In certain cases when closing the scene the selection in the subject hierarchy tree views was restored to an item that did not exist any more due to closing the scene. Since scene close is a batch processing event, selection restoration happened in that case as well, which now is fixed not to happen.