ENH: Allow setting node type display name for abstract node classes
The MRML scene provided node type display name for all registered nodes, but no type display name was available for abstract base classes.
For example, when a node selector was set to display vtkMRMLVolumeNode or vtkMRMLMarkupsNode then the display name was just "node" instead of "Volume" or "Markup", resulting in none display text as "Select a node" instead of "Select a Volume".

To fix this, RegisterAbstractNodeClass(std::string className, std::string typeDisplayName) method was added that allows specifying display name for each abstract node class.
This is now use for volume nodes (replacing a "special case" hack in qMRMLNodeComboBox) and markups nodes (replacing the hack of vtkMRMLMarkupsNode being a non-abstract class).

Co-authored-by: James Butler <jbutler@sonovol.com>