ENH: Use markups ROI by default in Crop Volume and Volume Rendering modules
Annotation ROIs are being phased out. As part of this effort, markups ROI is created by default in Crop Volume and Volume Rendering modules.

Also improved Volume Rendering module: ROI node is only added to the scene when needed (when ROI node display or cropping is requested), to avoid cluttering the scene with unused nodes.