BUG: Make "specify geometry" popup in Segment Editor set the exact extent
When the "Specify geometry" button was used in Segment Editor then the segmentation was never cropped,
but the extent of the segmentation was increased as much as needed to include the entire segment.

This led to unexpected results, as the segmentation did not end up with the requested geometry.
For example, the output segmentation size that was displayed in the specify geometry popup
was not correct.

Fixed the issue by forcing the output segmentation to have the exact extent that the user requested.

vtkSlicerSegmentationGeometryLogic default behavior did not change, but a new PadOutputGeometry
flag was added to the class, which is set to true by default.
The qMRMLSegmentationGeometryWidget sets PadOutputGeometry to false to enforce the requested geometry.