BUG: Use correct scene to get writers when exporting
This bug affects export and it was was introduced
when export functionality was first introduced in
c4a6881ed339b54dd6b86c2b8e8a0cea068988ea. For export,
qSlicerCoreIOManager::saveNodes needs to be called with
a temporary scene passed in. When qSlicerCoreIOManager::saveNodes
calls qSlicerCoreIOManagerPrivate::writers, the latter function
was still hard coded to work with the main scene. This is now fixed.

The effect of this bug is subtle to see while testing, because test cases often
end up having node IDs coincidentally match between the temporary scene and the main
scene. In the case of a mismatch, we were getting export failure.