BUG: Fix truncation of Segmentations when saving at reference geometry
Previously, saving a Segmentation with "Crop to minimum extent" disabled, the segmentation would be cropped to exactly the reference geometry extent. This could result in the segment being truncated if the reference geometry was smaller than the effective extent of the segmentation.

Fixed by saving the segmentation using an extent that contains both the reference and effective extents.
The same behavior is fixed when exporting segments to labelmap nodes.

Fix #6102
See discussion: https://discourse.slicer.org/t/editing-segmentation-of-model-and-saving-it/21338