ENH: Add warning messages when segmentation could be cropped
When changing the segmentation geometry, or when exporting a segmentation to file/node using a reference volume, it is possible to crop the segmentation.
This may result in a loss of data if the cropped region included part of the segment.

This commit adds warning messages when user attempts an operation that will crop the segmentation. The messages will only appear if the effective extent of the segmentation is not fully contained in the reference geometry. Also adds a check box to qMRMLSegmentationGeometryDialog that allows the reference geometry to be set without cropping the segmentation.