BUG: Fix display of point scalars in markups
Setting of scalar attribute and enabling scalar visibility had to be done in a markups nodes in a certain order, otherwise the coloring was note applied.

For example, without uncommenting the last two lines, the coloring did not appear:

iron = slicer.util.getFirstNodeByName("Iron")
curveNode = slicer.mrmlScene.GetFirstNodeByName("OC")
dispNode = curveNode.GetDisplayNode()
dispNode.SetAndObserveColorNodeID(iron.GetID())
dispNode.SetActiveScalar("PedigreeIDs", 0)
dispNode.SetScalarVisibility(1)
#dispNode.UpdateAssignedAttribute()
#dispNode.Modified()

The root cause of the problem was that in markups, when scalar visibility is changed then a filter must be replaced in the processing pipeline (by calling markupsNode->UpdateAssignedAttribute()).

Fixes the error reported here: https://discourse.slicer.org/t/changes-applied-in-a-python-script-propagate-only-to-the-gui-but-dont-get-displayed/21519/2